unify both tab implimentations
